requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pandas==1.3.3
scikit-learn==0.24.2
python-dotenv==0.19.0
requests==2.26.0 